Fixed unkillable node
Issue was really important. I was running an infinite while loop in the constructor, meaning construciton never finished. Therefore rclcp::spin was never called. Current solution is creating a timer based callback function (500Hz) to ask for images. 500 is obv way to much and actually the value needed for the IMU but whatever. Upside is all the other pulishers can be handled over the same callback structure. nice. Downside is I'll have to figure out how multithreading callbacks works, but thats actually a thing in ros2. I've created an issue containing two links with an example and stuff. Should be doable. Once we have that we just need the image transform (just copy from zed open capture), the other callbacks, and the xml with the frame stuff. All not too hard. Then Zed2 is done and only Blockfeld and thermal left. Easy.