Finished callbacks
Finished the callback structure for all callbacks for the stereo camera. Currently two issues. First is when looking for the sensor no device is found (using the id from the camera). Second is its not multithreading (as far as I'm aware). Also I still need to transfrom and split the images, and figure out why the timestamps aren't updating (maybe transform from nanoseconds to seconds).