Fixed blickfeld issue
Okay so when you find_package it adds a bunch of variables like the _DIR and the _LIBRARY and _LIBRARIES. But for some reason when find_package(blickfeld_scanner ... REQUIRED full) it would not create the _LIBRARY. I don't fully understand it, but it probably means the compiler does not know where the .so file is. We can manually find it using find_library(BLICKFELD_LIBRARY NAME blickfeld-scanner) and then target_link_library(lidar_publisher ). Since we still find_package(blickfeld-scanner), it does all the config stuff including properly loading Protobuf, so no Protobuf errors.