Thermal camera stuff
I added the code for the thermal camera. Should work in theory, but I can't connect to the camera as of now, so no idea honestly. Fancy python code though, not just c++. Also renamed zed2_interfaces to zed_open_capture for consistency and moved the senxor package needed for the thermal camera into ilumo_sensors. (could theoretically be an external package). For consistency I might do this with the ze_open_capture, but since it works atm probably I won't