Fixed senxor and started camera fix
So senxor needs to be in a different package otherwise it doesn't get build. Although maybe if I were to name the folder scripts instead of senxor but whatever. This works fine. Regarding the stereocamera, I have the hunch that the videoCap and sensCap objects are for some reason destroyed when the publisher initialization finishes. Therefore no images. Sidenode: Multithreading does not work currently.