Shared pointer capture
Creating shared pointer to adress the capture objects works. Otherwise the objects get destructed at the and of initialization of the publisher, but is still callable. So no error is raised but no data returned either. Data is still weird and I'm unsure on how to pass the synchronization but its a start.