Working on IMU messages
Kinda tricky. I created a burst class but import is failing. Also unsure how to best fill the data sequence. I could just created a vector and use pushback, but that copies the element every time. Allocating the space beforehand like for the Point Cloud data. Probably best but that will be hard af. We'll see.