Video running
So this has something to do with the sensor stream. When that is enabled, video does not work anymore, but sensor does. Something must be overwritten internally. Maybe its also a cmake issue. Also I now decalre the SENSOR_MOD_AVAILABLE variable and the VIDEO_MOD_AVAILABLE variable in the videcapture.hpp and the sensorcapture.hpp. This should actually be set in the cmake file, but if I do it like that it doesn't work. Regarding sensor stream, I left all code in and put it in a #ifdef SENSOR_STREAM. So if I wanna enable it, all I have to do is #define SENSOR_STREAM. Works for now. Camera stream now works HD at (almost) 60 FPS (sometimes drops to 30).